client_UI/requirements.txt
numpy
torch
torchvision
matplotlib
tensorboard
terminaltables
pillow
tqdm
imgaug